requirements.txt
allure-pytest==2.8.32
allure-python-commons==2.8.32
attrs==21.2.0
cffi==1.15.0
cryptography==36.0.2
faker==13.3.1
pytest==7.0.1
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
pymysql==1.0.2
pycparser==2.21
pytest-xdist==2.5.0
requests==2.22.0
selenium==4.0.0a5
SQLAlchemy==1.4.35
webdriver-manager==3.3.0
Appium-Python-Client==1.1.0